+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix checkpoint updates
+- Improve error handling for asset connectors
+
+
 ## 1.20.0 - 2025-09-09
 
 ### Changed